Dogelana Support, plus a new app icon.

+++ Assets/Samples/Solana SDK/0.0.6/Simple Wallet/Solana Wallet/Scripts/example/simple_screen_manager/utility/TokenItem.cs
@@ -43,7 +43,7 @@ namespace Solana.Unity.SDK.Example
         {
             _parentScreen = screen;
             TokenAccount = tokenAccount;
-            if (nftData != null && int.Parse(tokenAccount.Account.Data.Parsed.Info.TokenAmount.Amount) == 1)
+            if (nftData != null && long.Parse(tokenAccount.Account.Data.Parsed.Info.TokenAmount.Amount) == 1)
             {
                 await UniTask.SwitchToMainThread();
                 _nft = nftData;
